pyproject.toml
[project]
name = "hw1"
version = "1.0"
description = "BGU CS Course 'NLP with LLMs' - Spring 2025 - Michael Elhadad - HW1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "evaluate>=0.4.3",
    "google-genai>=1.10.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "nltk>=3.9.1",
    "openai>=1.75.0",
    "pyconll>=3.2.0",
    "pydantic>=2.11.3",
    "scikit-learn>=1.6.1",
]
